Ignore webhooks for unknown repos or events
This makes it respond usefully to GitHub's ping event. I also recently
learned that there are non-book repos in the GITenberg organization, so
ignore those.

This also revealed that pip was grabbing the wrong version of gitberg,
so fix that. I had only tested "setuptools gets the right version" and
"pip gets some version", not "pip gets the right version" before...