Add metadata automatically, and push it all to the website
This should be the last piece of adding metadata to books which lack it,
and then pushing the result to the website (including covers).

It also builds the epubs like Travis, although it doesn't actually make
GitHub releases because Travis is already doing that part.

https://github.com/GITenberg/The-Other-Now_51112 is a repository I
tested this on, and it seems to work nicely.