long-needed optimization of the worker
changes to make things work.
will be retiring the current EB-based autoupdate server in favor of a spun-up server because of minimal activity in the repos; a different architecture will work better for PG.

Couldn't get the worker running on Amazon Linux, so needed to start that from scratch, so may as well rething the architecture.